moved decodeMessage in utils
